+++ core/keystore/src/lib.rs
@@ -86,7 +86,6 @@ impl Store {
 
 	/// Generate a new key, placing it into the store.
 	pub fn generate<TPair: Pair>(&self, password: &str) -> Result<TPair> {
-		println!("==== generating a new key inside keystore, password: {}", password);
 		let (pair, phrase, _) = TPair::generate_with_phrase(Some(password));
 		let mut file = File::create(self.key_file_path::<TPair>(&pair.public()))?;
 		::serde_json::to_writer(&file, &phrase)?;
@@ -96,7 +95,6 @@ impl Store {
 
 	/// Create a new key from seed. Do not place it into the store.
 	pub fn generate_from_seed<TPair: Pair>(&mut self, seed: &str) -> Result<TPair> {
-		println!("==== generating a new key inside keystore from seed: {}", seed);
 		let pair = TPair::from_string(seed, None)
 			.ok().ok_or(Error::InvalidSeed)?;
 		self.insert_pair(&pair);
@@ -152,38 +150,6 @@ impl Store {
 
 		Ok(public_keys)
 	}
-
-	// // TODO: This function duplicates most of the logic of contents(). Can they share logic?
-	// pub fn contents_by_type_id(&self, id: KeyTypeId) -> Result<Vec<Vec<u8>>> {
-	// 	let mut public_keys: Vec<Vec<u8>> = self.additional.keys()
-	// 		.filter_map(|(ty, public)| {
-	// 			if *ty != id {
-	// 				return None
-	// 			}
-	// 			Some(public.to_vec())
-	// 		})
-	// 		.collect();
-
-	// 	let key_type: [u8; 4] = id.to_le_bytes();
-	// 	for entry in fs::read_dir(&self.path)? {
-	// 		let entry = entry?;
-	// 		let path = entry.path();
-
-	// 		// skip directories and non-unicode file names (hex is unicode)
-	// 		if let Some(name) = path.file_name().and_then(|n| n.to_str()) {
-	// 			match hex::decode(name) {
-	// 				Ok(ref hex) => {
-	// 					if hex[0..4] != key_type { continue	}
-	// 					let public = &hex[4..];
-	// 					public_keys.push(public.to_vec());
-	// 				}
-	// 				_ => continue,
-	// 			}
-	// 		}
-	// 	}
-
-	// 	Ok(public_keys)
-	// }
 
 	fn key_file_path<TPair: Pair>(&self, public: &TPair::Public) -> PathBuf {
 		let mut buf = self.path.clone();
